requirements.txt
flask
arrow
webview
colorama
requests
LnkParse3
python-dotenv
beautifulsoup4